rename service name
